fix(plugin-python): send and read bigint as a digit string
bigint mapped to a bare int, so model_dump(mode="json") put a JSON number on
the wire. A ContractKit server's bigint schema only accepts a string ("123",
or the TypeScript SDK's "123n"): ServerKit's body parser revives "123n", the
Zod preprocess converts any other string, and a number is rejected with
"expected bigint, received number". Reading a "123n" back failed against int.

A new _scalars.py module, emitted with the base client, defines
BigInt = Annotated[int, BeforeValidator(...), PlainSerializer(str, when_used=
"json")]. It reads a digit string, "123n" or a number, and writes a plain digit
string, which is what the Kotlin, Swift and C# SDKs write. Models import it,
and the client imports it wherever a body, query param or response type
carries a bigint. Relative imports now render after the stdlib ones.

No generated server can put a bigint in a response today: the Koa and Fastify
routers leave serialization to JSON.stringify, which throws on a BigInt. The
reader accepts every form so it keeps working whichever way that is fixed.

+++ packages/output-tests/tests/__snapshots__/python/pysdk/_models_kitchen.py
@@ -6,6 +6,7 @@ from uuid import UUID
 from typing import Annotated, Any, Literal
 from pydantic import BaseModel, ConfigDict, Field
 from decimal import Decimal
+from ._scalars import BigInt
 
 # A named enum, so a field default has to resolve to a member rather than its wire spelling
 Rating = Literal["good", "neutral", "bad"]
@@ -89,7 +90,7 @@ class Folder(BaseModel):
     pinned: Doc | Folder | None
     instrument: Card | Bank | None = None
     origin: dict[str, Any] | None = None
-    size: int
+    size: BigInt
     price: Decimal
     day: date | None = None
     at: time | None = None
